requirements.txt
streamlit
scikit-learn
tensorflow
keras
pillow
numpy
streamlit-option-menu
mysql
mysql-connector